requirements.txt
fastapi
uvicorn
python-dotenv
passlib[jwt]
jose
jwt
pydantic[email]
python-jose
python-multipart
json